+++ mobile/android/geckoview_example/src/main/java/org/mozilla/geckoview_example/GeckoViewActivity.java
@@ -1655,6 +1655,14 @@ public class GeckoViewActivity
             }
             getWindow().setAttributes(layoutParams);
         }
+
+        @Override
+        public void onShowDynamicToolbar(final GeckoSession session) {
+            final View toolbar = findViewById(R.id.toolbar);
+            if (toolbar != null) {
+                toolbar.setTranslationY(0f);
+            }
+        }
     }
 
     private class ExampleProgressDelegate implements GeckoSession.ProgressDelegate {