+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/doc-files/CHANGELOG.md
@@ -17,8 +17,11 @@ exclude: true
 - Added [`GeckoSession.PermissionDelegate.PERMISSION_STORAGE_ACCESS`][92.1] to
   control the allowing of third-party frames to access first-party cookies and
   storage. ([bug 1543720]({{bugzilla}}1543720))
+- Added [`ContentDelegate#onShowDynamicToolbar(GeckoSession)`][92.2] to notify
+  the app that it must fully-expand its dynamic toolbar ([bug 1690296]({{bugzilla}}1690296)).
 
 [92.1]: {{javadoc_uri}}/GeckoSession.PermissionDelegate.html#PERMISSION_STORAGE_ACCESS
+[92.2]: {{javadoc_uri}}/GeckoSession.ContentDelegate.html#onShowDynamicToolbar-org.mozilla.geckoview.GeckoSession-
 
 ## v91
 - Extended [`Autocomplete`][78.7] API to support addresses.
@@ -1029,4 +1032,4 @@ to allow adding gecko profiler markers.
 [65.24]: {{javadoc_uri}}/CrashReporter.html#sendCrashReport-android.content.Context-android.os.Bundle-java.lang.String-
 [65.25]: {{javadoc_uri}}/GeckoResult.html
 
-[api-version]: 6cc809570a2e6e41a82ba612eee0f3059aecb14d
+[api-version]: b1c56ee140a5ecc48c10082555fd6b3e96b3773e